OrderService/requirement.txt
aniso8601==9.0.1
asgiref==3.5.2
blinker==1.5
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
flask-mongoengine==1.0.0
flask-requests==0.0.14
Flask-RESTful==0.3.9
Flask-WTF==1.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.2
pampy==0.3.0
pymongo==4.2.0
pytz==2022.5
six==1.16.0
Werkzeug==2.2.2
WTForms==3.0.1
